+++ Assets/Sprites/RedRunner/Tiles Atlas.spriteatlas
@@ -47,7 +47,7 @@ SpriteAtlas:
     bindAsDefault: 1
     storedHash:
       serializedVersion: 2
-      Hash: 7a96b2cb0b7189353533c3ad7753ca2e
+      Hash: d6950cb28b4092cca5880df168d8dcc8
   m_MasterAtlas: {fileID: 0}
   m_PackedSprites:
   - {fileID: 21300000, guid: d0fc8af013ce8c8439e4f006b8ca3f81, type: 3}
